# core-common-contract-mgmt-sdk/src/main/resources/api-spec/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "core-common-contract-mgmt",
    "description": "Contract Management Core Application",
    "contact": {
      "name": "Firefly Software Foundation",
      "email": "dev@getfirefly.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://core.getfirefly.io/contracts",
      "description": "Development Environment"
    },
    {
      "url": "/",
      "description": "Local Development Environment"
    }
  ],
...
  "paths": {
...
    "/api/v1/contracts/{contractId}/documents/{contractDocumentId}": {
      "get": {
        "tags": [
          "Contract Documents"
        ],
        "summary": "Get contract document by ID",
        "description": "Returns a contract document based on its ID",
        "operationId": "getContractDocumentById",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "contractDocumentId",
            "in": "path",
            "description": "ID of the contract document to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Contract document not found"
          },
          "200": {
            "description": "Successfully retrieved contract document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDocumentDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Contract Documents"
        ],
        "summary": "Update contract document",
        "description": "Updates an existing contract document with the provided information",
        "operationId": "updateContractDocument",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "contractDocumentId",
            "in": "path",
            "description": "ID of the contract document to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractDocumentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Contract document not found"
          },
          "200": {
            "description": "Contract document successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDocumentDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid contract document data provided"
          }
        }
      },
      "delete": {
        "tags": [
          "Contract Documents"
        ],
        "summary": "Delete contract document",
        "description": "Deletes a contract document based on its ID",
        "operationId": "deleteContractDocument",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "contractDocumentId",
            "in": "path",
            "description": "ID of the contract document to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "204": {
            "description": "Contract document successfully deleted"
          },
          "404": {
            "description": "Contract document not found"
          }
        }
      }
    },
    "/api/v1/contracts/{contractId}": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get contract by ID",
        "description": "Returns a contract based on its ID",
        "operationId": "getContractById",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contract",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDTO"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found"
          }
        }
      },
      "put": {
        "tags": [
          "Contracts"
        ],
        "summary": "Update contract",
        "description": "Updates an existing contract with the provided information",
        "operationId": "updateContract",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid contract data provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Contract successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDTO"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Contracts"
        ],
        "summary": "Delete contract",
        "description": "Deletes a contract based on its ID",
        "operationId": "deleteContract",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Contract not found"
          },
          "204": {
            "description": "Contract successfully deleted"
          }
        }
      }
    },
...
    "/api/v1/contracts/{contractId}/documents/filter": {
      "post": {
        "tags": [
          "Contract Documents"
        ],
        "summary": "Filter contract documents",
        "description": "Returns a paginated list of contract documents based on filter criteria",
        "operationId": "filterContractDocuments",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestContractDocumentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid filter criteria provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contract documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contractId}/documents": {
      "post": {
        "tags": [
          "Contract Documents"
        ],
        "summary": "Create a new contract document",
        "description": "Creates a new contract document with the provided information",
        "operationId": "createContractDocument",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractDocumentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Contract not found"
          },
          "201": {
            "description": "Contract document successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDocumentDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid contract document data provided"
          }
        }
      }
    },
    "/api/v1/contracts/filter": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Filter contracts",
        "description": "Returns a paginated list of contracts based on filter criteria",
        "operationId": "filterContracts",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestContractDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid filter criteria provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contracts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContractDocumentDTO": {
        "required": [
          "contractId",
          "documentId",
          "documentTypeId"
        ],
        "type": "object",
        "properties": {
          "contractDocumentId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "contractId": {
            "type": "string",
            "format": "uuid"
          },
          "documentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
      "ContractDTO": {
        "required": [
          "contractStatus"
        ],
        "type": "object",
        "properties": {
          "contractId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "contractNumber": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "contractStatus": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_APPROVAL",
              "ACTIVE",
              "SUSPENDED",
              "TERMINATED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "productCatalogId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "FilterOptions": {
        "type": "object",
        "properties": {
          "caseInsensitiveStrings": {
            "type": "boolean"
          },
          "includeInheritedFields": {
            "type": "boolean"
          }
        },
        "description": "Filter options for customizing filter behavior"
      },
...
      "PaginationRequest": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "description": "The zero-based page number to retrieve.",
            "format": "int32",
            "example": 0,
            "default": 0
          },
          "pageSize": {
            "type": "integer",
            "description": "The number of items per page.",
            "format": "int32",
            "example": 10,
            "default": 10
          },
          "sortBy": {
            "type": "string",
            "description": "The field to sort the results by.",
            "example": "name"
          },
          "sortDirection": {
            "type": "string",
            "description": "The direction of sorting, either ASC or DESC.",
            "example": "DESC",
            "default": "DESC"
          }
        },
        "description": "Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction."
      },
      "RangeFilter": {
        "type": "object",
        "properties": {
          "ranges": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/RangeObject"
            },
            "description": "Map of field names to their range criteria"
          }
        },
        "description": "Represents range filters for numeric fields"
      },
      "RangeObject": {
        "type": "object",
        "properties": {
          "from": {
            "type": "object",
            "description": "Start value of the range"
          },
          "to": {
            "type": "object",
            "description": "End value of the range"
          }
        },
        "description": "Represents a range with from and to values"
      },
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "The list of items for the current page.",
            "items": {
              "type": "object",
              "description": "The list of items for the current page."
            }
          },
          "totalElements": {
            "type": "integer",
            "description": "The total number of elements across all pages.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages based on the data size and page size.",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "description": "The current page number, typically zero-based.",
            "format": "int32"
          }
        },
        "description": "Represents a paginated response containing a list of items and pagination metadata."
      },
...
      "FilterRequestContractDocumentDTO": {
        "required": [
          "pagination"
        ],
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ContractDocumentDTO"
          },
          "rangeFilters": {
            "$ref": "#/components/schemas/RangeFilter"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequest"
          },
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          }
        },
        "description": "Generic filter request that includes both filter criteria and pagination"
      },
      "FilterRequestContractDTO": {
        "required": [
          "pagination"
        ],
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ContractDTO"
          },
          "rangeFilters": {
            "$ref": "#/components/schemas/RangeFilter"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequest"
          },
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          }
        },
        "description": "Generic filter request that includes both filter criteria and pagination"
      },
...
    }
  }
}